tmf.ui: Add a loadingTrace method to time graph view
This method is called when a trace is being loaded, before the entries
are built or the view updated. Implementing classes that have specific
trace information to keep can thus override this method when a trace is
selected or opened.

Use case: the base data provider view (supporting multiple data
providers) have a specific set of entries/data provider to keep for each
trace.

Change-Id: I8628d6d857979607015214d96e9b1a7793fdc258
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>