graph: Add an implementation of the graph on disk
This implementation uses an history-tree like structure. The history
tree classes HtIo, HTNode and HistoryTree have been copied from the
segment store and datastore and modified to remove generics (which
improves performances by at least 3 to 5 times) and implement specific
methods for query in the tree.

Experimentation was done to use available backends like the state system,
but it does not support 0-length states (for vertical edges) or segment
stores, which could have worked, but generics and query patterns made
the performances really bad for the graph.

Performance-wise, the graph construction remains slightly similar to
what it was, with a different less than 5% sometimes, within the margin
of errors. Building the critical path though takes a lot longer, as each
edge to be queried on disk. Is there room for improvement in the
queries? Probably! But this is good enough. For small traces, it remains
within one second. But the main improvement is for larger traces, that
used to either throw OOME or took close to the whole RAM, making Trace
Compass close to unresponsive. Now, large traces do not OOME and the
interface remains responsive.

Benchmark results (before => after):

Trace                Build graph (sec)   Build critical path (sec)
------------------------------------------------------------------
Trace2              1.91  =>   1.87          0.008  =>   0.047
Multi-thread        1.58  =>   1.58          0.031  =>   1.31
Django              2.82  =>   2.87          0.003  =>   0.034
OSEvents            9.12  =>   8.91          0.035  =>   0.409
Large trace 1*     30.46  =>  28.88          0.513  =>  12.62
Large trace 2*     57.88  =>  55.42          0.127  =>   2.09
Very large**        OOME  => 750.0            --    => 134.4

* Arbitrary large traces that previously worked in Trace Compass, but when
calculating the execution graph and critical path, used to bring the UI
to the brink of unresponsiveness.

** 7.4GB kernel trace with only required events and system call of ~15
minutes normal developer laptop usage.

Change-Id: Ib2fda449da266830c4d847d5c3f4cb11fc8bd283
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>